Remove release profile
The release profile is not needed given this is a simple, single module
project and can compile fine on any supported operating system (Windows,
OS X, Linux).

Removing the release profile causes the JavaDoc plugin to execute in a
more common build lifecycle, leading to earlier identification of
JavaDoc errors.