Skip Checkstyle until acegi-standard-project 0.0.3
This skip only applies to AppVeyor, as it's a Windows build.

The root cause of the issue is the NewlineAtEndOfFile check uses the
default "system" value for "lineSeparator". In upstream
acegi-standard-project commit b81a0a0823b8c9a1a7c0ff07810c2bf2e395f241 I
have explicitly set to use Unix line endings, but it's not worth making
a 0.0.3 release on account of this single issue.