Tests to more carefully close Env where possible
The purpose of this change is to reduce disk space consumption during
tests, most critically on Windows (which does not delete the JUnit @Rule
TemporaryFolder-created paths unless Env is formally closed; this is
likely a platform-specific difference between Linux and Windows as Linux
removes these files without needing the Env to be formally closed).

Some corner cases still do not formally close Env, such as when a
@Test(expected=Throwable.class) is used, plus the TutorialTest does not
close Env given it complicates the tutorial for a platform-specific,
testing-related reason only.