Preliminary KeyRange and tests (#7)
This commit captures the available key ranges and unit tests that
confirm the JavaDoc specification is correct. The KeyRange class
provides package protected methods that encapsulate the iteration and
control logic such that the eventual Iterator class will not have any
awareness of the various KeyRangeTypes and it should be possible to
add new KeyRangeTypes without modifying the actual Iterator.